+++ requirements.txt
@@ -8,7 +8,7 @@ MarkupSafe==0.23
 numpy==1.9.1
 python-dateutil==2.6.0
 requests==2.12.4
-s3transfer==0.1.10
 six==1.10.0
 wheel==0.29.0
+python-dotenv==0.6.2
 git+git://github.com/gipit/gippy.git@develop
